Improve clone for COW SvPV when close to RefCnt Max
With the previous change one COWed SvPV was having
a CowRefcnt value set to the maximum value
( SV_COW_REFCNT_MAX - 1 or 254), the cloned SvPV was not COW
anymore.

This commit add an improved support for such SvPVs and preserve
the COW status on these strings, so (massive) cloning of the
same SvPV will improve memory usage and reduce duplicates.