Handle the COW SvPV with LEN=0 differently
The first implementation of COW cloning had
one issue with SvPV COW having LEN=0. (SEGV)

With this commit we are now treating them the
same as a new COW PV when doing the overflow
on the CowRefcnt.

In other word, we clone the whole PV, but set the
COW flag and CowRefcnt to 0.